+++ src/ann/xtensor_lib.cpp
@@ -6,7 +6,7 @@
 #include "ann/xtensor_lib.h"
 
 
-string shape2str(xt::svector<unsigned long> vec){
+string shape2str(xt::svector<size_t> vec){ // temperary change to size_t
     stringstream ss;
     ss << "(";
     for(int idx=0; idx < vec.size(); idx++){
